+++ tpl/runtime.cjs.js
@@ -1,8 +1,8 @@
 /* eslint-disable */
 let mocker;
 
-/* __hrf_env_if__ */if (process.env.__hrm_enviroment_key__ === '__hrm_enviroment_val__') {
-  const HttpRequestMock = require('http-request-mock');
+/* __hrf_env_if__ */if (process.env.__hrm_environment_key__ === "__hrm_environment_val__") {
+  const HttpRequestMock = require("http-request-mock");
   mocker = HttpRequestMock.setup(__hrm_proxy_server__);
   __hrm_mock_items__
 /* __hrf_env_if__ */}