+++ tpl/samples/dynamic.js
@@ -4,14 +4,14 @@
  * @method get
  */
 /* eslint-disable */
-const faker = require('http-request-mock/plugin/faker.js');
+const faker = require("http-request-mock/plugin/faker.js");
 module.exports = (request) => {
   return  {
     ret: 0,
-    msg: 'ok',
+    msg: "ok",
     name: faker.name(),
     email: faker.email(),
-    phone: faker.phone('1-###-###-####'),
+    phone: faker.phone("1-###-###-####"),
     url: request.url,
   };
 }